+++ src/muncher/classes.js
@@ -2,6 +2,7 @@
 import { getClasses } from "./classes/classes.js";
 import { munchNote, getCampaignId, download } from "./utils.js";
 import { getSubClasses } from "./classes/subclasses.js";
+import { getClassOptions } from "./classes/options.js";
 
 function getSubClassesData(className) {
   const cobaltCookie = game.settings.get("ddb-importer", "cobalt-cookie");
@@ -31,6 +32,39 @@ function getSubClassesData(className) {
         return data;
       })
       .then((data) => getSubClasses(data.data))
+      .then((data) => resolve(data))
+      .catch((error) => reject(error));
+  });
+}
+
+function getClassOptionsData(className) {
+  const cobaltCookie = game.settings.get("ddb-importer", "cobalt-cookie");
+  const campaignId = getCampaignId();
+  const parsingApi = game.settings.get("ddb-importer", "api-endpoint");
+  const betaKey = game.settings.get("ddb-importer", "beta-key");
+  const body = { cobalt: cobaltCookie, campaignId: campaignId, betaKey: betaKey, className: className };
+  const debugJson = game.settings.get("ddb-importer", "debug-json");
+
+  return new Promise((resolve, reject) => {
+    fetch(`${parsingApi}/proxy/classes/options`, {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify(body),
+    })
+      .then((response) => response.json())
+      .then((data) => {
+        if (debugJson) {
+          download(JSON.stringify(data), `subclass-${className}-raw.json`, "application/json");
+        }
+        if (!data.success) {
+          munchNote(`Failure: ${data.message}`);
+          reject(data.message);
+        }
+        return data;
+      })
+      .then((data) => getClassOptions(data.data, className))
       .then((data) => resolve(data))
       .catch((error) => reject(error));
   });
@@ -72,6 +106,7 @@ function getClassesData() {
 export async function parseClasses() {
   const classesResults = await getClassesData();
 
+ // const classesResults = [];
   const subClassResults = await Promise.allSettled([
     getSubClassesData("Cleric"),
     getSubClassesData("Druid"),
@@ -86,9 +121,27 @@ export async function parseClasses() {
     getSubClassesData("Artificer"),
     getSubClassesData("Rogue"),
     getSubClassesData("Monk"),
+    getSubClassesData("Blood Hunter"),
   ]);
 
-  return classesResults.concat(subClassResults);
+  const classOptionsResults = await Promise.allSettled([
+    getClassOptionsData("Cleric"),
+    getClassOptionsData("Druid"),
+    getClassOptionsData("Sorcerer"),
+    getClassOptionsData("Warlock"),
+    getClassOptionsData("Wizard"),
+    getClassOptionsData("Paladin"),
+    getClassOptionsData("Ranger"),
+    getClassOptionsData("Bard"),
+    getClassOptionsData("Barbarian"),
+    getClassOptionsData("Fighter"),
+    getClassOptionsData("Artificer"),
+    getClassOptionsData("Rogue"),
+    getClassOptionsData("Monk"),
+    getClassOptionsData("Blood Hunter"),
+  ]);
+
+  return classesResults.concat(subClassResults, classOptionsResults);
 }