+++ src/muncher/classes/shared.js
@@ -21,6 +21,43 @@ const CLASS_TEMPLATE = {
   "img": null
 };
 
+export const NO_TRAITS = [
+  "Speed",
+  "Ability Score Increase",
+  "Ability Score Improvement",
+  "Size",
+  "Feat",
+  "Languages",
+  "Hit Points",
+  "Proficiencies",
+  "Bonus Proficiencies",
+  "Bonus Proficiency",
+  "Extra Attack",
+  "Tool Proficiency",
+  "Bonus Cantrip",
+  "Bonus Cantrips",
+];
+
+export const FEATURE_DUP = [
+  "Expertise",
+  "Fighting Style",
+  "Land's Stride",
+  "Otherworldly Patron",
+  "Pact Magic",
+  "Potent Spellcasting",
+  "Timeless Body",
+  "Unarmored Defense",
+  "Circle Spells",
+  "Divine Strike",
+  "Evasion",
+  "Channel Divinity",
+  "Expanded Spell List",
+  "Oath Spells",
+  "Psionic Power",
+  "Psychic Blades",
+  "Spellcasting",
+];
+
 function buildBase(data) {
   let result = JSON.parse(JSON.stringify(CLASS_TEMPLATE));
 
@@ -44,6 +81,9 @@ export function getClassFeature(feature, klass, subClassName = "") {
   logger.debug("Class feature build started");
 
   let result = buildBase(feature);
+
+  const duplicateFeature = FEATURE_DUP.includes(feature.name);
+  result.name = (duplicateFeature) ? `${feature.name} (${klass.name})` : feature.name;
 
   result.flags.ddbimporter['featureId'] = feature.id;
   result.flags.ddbimporter['requiredLevel'] = feature.requiredLevel;